Fix (v.1.27) : Forzar el tipo de Días de Cobertura a número entero nativo en la base de datos (Models & Schemas) y eliminar los decimales sobrantes en la exportación a Excel

+++ app/models.py
@@ -97,7 +97,7 @@ class ProductoMetricas(Base):
     abc = Column(String)
     xyz = Column(String)
     matriz_abc = Column(String)
-    dias_cobertura = Column(Float)
+    dias_cobertura = Column(Integer)
     riesgo_rotura = Column(Boolean, default=False)
     
     producto = relationship("Producto")
